🔧 Fix: Resolución completa de 46 errores de compilación críticos
Corrige errores de tipos en conversiones Entity/DTO, BigDecimal/Double y String/Categoria:

## Categorías de errores corregidos:

### 1. BigDecimal → Double (10 errores)
- ReporteControlador.java: calcularTotalVentas, calcularTicketPromedio
- DashboardControladorAPI.java: métricas de ventas y tickets
- ExportacionServicioImpl.java: cálculos financieros
- ProductoServicioImpl.java: conversión de precio
- HistorialVentasControlador.java: agregación de totales con map(BigDecimal::valueOf)
- DashboardControladorVista.java: métricas comparativas

### 2. Entity ↔ DTO conversiones individuales (~10 errores)
- ClienteImportacionServicioImpl: Cliente ↔ ClienteDTO con modelMapper
- ValidacionServicioImpl: ProductoDTO para validaciones de stock
- ProductoImportacionServicioImpl: gestión de productos existentes
- ImportacionServicioImpl: conversión para persistencia
- ValidacionDatosServicioImpl: validaciones de referencias

### 3. List<DTO> ↔ List<Entity> conversiones (~15 errores)
- ReporteClienteServicioImpl: distribución de antigüedad
- ClienteImportacionServicioImpl: búsqueda por RUT
- ReporteClienteControlador: clientes inactivos
- ValidacionDatosServicioImpl: validación de módulos
- ImportacionExportacionControlador: exportaciones (productos/ventas)
- ExportacionServicioImpl: obtención de datos para reportes

### 4. String → Categoria conversiones (2 errores)
- ProductoServicioImpl: inyección de CategoriaRepositorio con findByNombre
- ReporteExportacionControlador: uso de constructor DTO directo

### 5. Inferencia de tipos genéricos (3 errores)
- ValidacionDatosServicioImpl: validacionRapida con conversiones DTO→Entity

## Mejoras técnicas:
- Inyección de ModelMapper en 10+ servicios
- Uso consistente de streams con map/collect para conversiones masivas
- Inyección de CategoriaRepositorio donde necesario
- Uso de Optional.ifPresent para conversiones seguras

+++ src/main/java/informviva/gest/controlador/DashboardControladorVista.java
@@ -50,8 +50,8 @@ public class DashboardControladorVista {
 
         // KPIs de la semana actual
         MetricaDTO ventasMetrica = calcularMetrica(
-                ventaServicio.calcularTotalVentas(semanaActual[0], semanaActual[1]),
-                ventaServicio.calcularTotalVentas(semanaAnterior[0], semanaAnterior[1])
+                ventaServicio.calcularTotalVentas(semanaActual[0], semanaActual[1]).doubleValue(),
+                ventaServicio.calcularTotalVentas(semanaAnterior[0], semanaAnterior[1]).doubleValue()
         );
 
         MetricaDTO transaccionesMetrica = calcularMetrica(
@@ -60,8 +60,8 @@ public class DashboardControladorVista {
         );
 
         MetricaDTO ticketMetrica = calcularMetrica(
-                ventaServicio.calcularTicketPromedio(semanaActual[0], semanaActual[1]),
-                ventaServicio.calcularTicketPromedio(semanaAnterior[0], semanaAnterior[1])
+                ventaServicio.calcularTicketPromedio(semanaActual[0], semanaActual[1]).doubleValue(),
+                ventaServicio.calcularTicketPromedio(semanaAnterior[0], semanaAnterior[1]).doubleValue()
         );
 
         MetricaDTO clientesMetrica = calcularMetrica(
